refactor: centralize unwrapListModel in shared model-utils
Five emitters (node, python, php, go, dotnet) carried byte-identical
copies of the list-wrapper unwrapping logic, so a fix in one would
silently drift from the others. Verified behavior-preserving by
byte-identical staging regeneration for all five languages.
